Attempt to keep main.js less cluttered
Added myCode.js, trying to keep main.js less cluttered but not working ATM.
