+++ index.html
@@ -7,9 +7,5 @@
   <body>
     <p>Hello World!</p>
     <h2>Current UTC Time:</h2>
-    <h3 id="hours"></h3>
-    <h3 id="minutes"></h3>
-    <h3 id="seconds"></h3>
-    <script type="text/javascript" src="main.js"></script>
   </body>
 </html>
