+++ main.js
@@ -2,6 +2,9 @@
 const {app, BrowserWindow} = require('electron');
 const path = require('path');
 
+//all of the code I wrote is in myCode.js, to make this file cleaner and have less junk
+import { runDate, test } from './myCode.js';
+
 function createWindow () {
   // Create the browser window.
   const mainWindow = new BrowserWindow({
@@ -24,10 +27,10 @@ function createWindow () {
 // Some APIs can only be used after this event occurs.
 app.whenReady().then(() => {
   console.log('running'); //debug
-  //createWindow();
+  createWindow();
   console.log('created window, running date function'); //debug
-  runDate();
-  //test();
+  runDate();  //function to run and test the whole thing
+  //test();   //testing function for small things, e.g. sleep() function
 
   app.on('activate', function () {
     // On macOS it's common to re-create a window in the app when the
@@ -42,60 +45,3 @@ app.whenReady().then(() => {
 app.on('window-all-closed', function () {
   if (process.platform !== 'darwin') app.quit()
 })
-
-// In this file you can include the rest of your app's specific main process
-// code. You can also put them in separate files and require them here.
-
-
-function runDate() {
-  let hours, minutes, seconds;
-  for (let hours = 0; hours < 24; hours++) {
-    hours = updateHours();
-    for(let minutes = 0; minutes < 60; minutes++) {
-      minutes = updateMinutes();
-      for(let seconds = 0; seconds < 60; seconds++) {
-        seconds = updateSeconds();
-        console.log("The current time is: " + hours + " " + minutes + " " + seconds); //for debug uses in the future, logs to console when running through npm start
-        //console.log(seconds.length, seconds);
-        sleep(20);
-      }
-    }
-  }
-}
-
-//these three functions do exactly their description, they update the hours, minutes and seconds on the clock
-function updateHours() {
-  let now = new Date();
-  let hours = String(now.getHours());
-  return hours;
-}
-function updateMinutes() {
-  let now = new Date();
-  let minutes = String(now.getMinutes());
-  return minutes;
-}
-function updateSeconds() {
-  let now = new Date();
-  let seconds = String(now.getSeconds());
-  if (seconds.length == 1) {
-    seconds = '0'+String(seconds);
-  }
-  return seconds;
-}
-
-//this function blocks any action until it finishes, basically making EVERYTHING wait for it to finish
-function sleep(milliseconds) {
-  const date = Date.now();
-  let currentDate = null;
-  do {
-    currentDate = Date.now();
-  } while (currentDate - date < milliseconds);
-
-//function for testing stuff, should not run in normal cycle
-function test() {
-  while (true) {
-    let now = new Date;
-    let milli = String(now.getTime());
-    console.log(milli);
-  }
-}
